chatbot/requirements.txt
openai
langchain
chromadb
python-dotenv
pandas
fastapi
uvicorn
tiktoken
